regular update

+++ accessory/parent_db.java
@@ -11,8 +11,8 @@ public abstract class parent_db
 	public abstract String sanitise_string(String input_);
 	public abstract ArrayList<HashMap<String, String>> execute(String source_, String type_, String[] cols_, HashMap<String, String> vals_, String where_, int max_rows_, String order_, HashMap<String, db_field> cols_info_);
 	public abstract HashMap<String, Object> get_data_type(String data_type_);
-	public abstract int get_default_size(String data_type_);
-	public abstract int get_max_size(String data_type_);
+	public abstract long get_default_size(String data_type_);
+	public abstract long get_max_size(String data_type_);
 	public abstract String get_value(String input_);
 	public abstract String get_variable(String input_);
